+++ src/mqtt-plus-base.ts
@@ -111,8 +111,10 @@ export class BaseTrait<T extends APISchema = APISchema> extends TraceTrait<T> {
             destroy: async (): Promise<void> => {
                 if (!this.onRequest.has(key))
                     throw new Error(`destroy: ${kind} "${name}" not registered`)
-                await spool.unroll(false)?.catch((err: Error) => {
-                    this.error(err, `destroy: failed to cleanup: ${err.message}`)
+                await spool.unroll(false)?.catch((err: unknown) => {
+                    const error = ensureError(err, `destroy: ${kind} "${name}" failed to cleanup`)
+                    this.error(error)
+                    throw error
                 })
             }
         }